added travis.yml
