+++ lib/main.dart
@@ -1,30 +1,20 @@
 import 'package:flutter/material.dart';
-import 'package:flutter_manga_reader/views/top.dart';
+import 'package:flutter_manga_reader/views/manga_list.dart';
 
 void main() {
   runApp(MyApp());
 }
 
 class MyApp extends StatelessWidget {
-  // This widget is the root of your application.
   @override
   Widget build(BuildContext context) {
     return MaterialApp(
       title: 'Flutter Demo',
       theme: ThemeData(
-          // This is the theme of your application.
-          //
-          // Try running your application with "flutter run". You'll see the
-          // application has a blue toolbar. Then, without quitting the app, try
-          // changing the primarySwatch below to Colors.green and then invoke
-          // "hot reload" (press "r" in the console where you ran "flutter run",
-          // or press Run > Flutter Hot Reload in IntelliJ). Notice that the
-          // counter didn't reset back to zero; the application is not restarted
           primaryColor: Colors.black,
           primaryColorDark: Colors.grey,
           accentColor: Colors.amber),
-      home: TopManga(title: 'Manga List'),
+      home: MangaListView(title: 'Manga List'),
     );
   }
 }
-
